move more icons into ide.icons
flatten UrlAsynchronousIcon into AsynchronousWorkerIcon
