+++ core/util/src/main/java/org/alice/ide/icons/AsynchronousIcon.java
@@ -40,7 +40,7 @@
  * THE USE OF OR OTHER DEALINGS WITH THE SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  *******************************************************************************/
-package edu.cmu.cs.dennisc.javax.swing;
+package org.alice.ide.icons;
 
 import edu.cmu.cs.dennisc.java.util.Lists;
